Forked code from I am a compsci student and hobby developer. I primarily use C++, but I've very recently picked up C#.
I've been working on a chess engine for a while, and somewhat recently I decided to move the front-end part of the chess engine code to C# because string parsing in C++ is a pain and I was already interested in learning C#

This is a console that takes a FEN record as input (https://www.chessprogramming.org/Forsyth-Edwards_Notation#Samples) and outputs an ASCII chess board. It also converts the string to a blittable struct for C#/C++ interoperability.

I'd like feedback on some C# best practices and .NET tricks to improve future code.

If there is anything wrong with the repository, it is because I have never used Github in my entire life

Project uses C# 10 and .NET 6.0, compiled with Visual Studio 2022